+++ src/Main/java/CAB302/Common/Trade.java
@@ -30,11 +30,11 @@ public class Trade extends BaseObject implements iGet, iList {
     public int getQuantity() { return this.quantity; }
     public void setQuantity(int quantity) { this.quantity = quantity; }
 
-    private double price;
+    private int price;
 
     @Column(name = "price")
-    public double getPrice() { return this.price; }
-    public void setPrice(double price) { this.price = price; }
+    public int getPrice() { return this.price; }
+    public void setPrice(int price) { this.price = price; }
 
     private Timestamp createdDate;
